combine user settings and account area into 1 tab in user area
