+++ cmd/frontend/internal/app/verify_email.go
@@ -49,7 +49,7 @@ func serveVerifyEmail(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	http.Redirect(w, r, "/user/account", http.StatusFound)
+	http.Redirect(w, r, "/user/settings/emails", http.StatusFound)
 }
 
 func httpLogAndError(w http.ResponseWriter, msg string, code int, errArgs ...interface{}) {
