+++ web/src/auth.ts
@@ -30,6 +30,7 @@ export function refreshAuthenticatedUser(): Observable<never> {
                 siteAdmin
                 tags
                 url
+                settingsURL
                 organizations {
                     nodes {
                         id
