+++ web/src/user/area/navitems.ts
@@ -1,5 +1,4 @@
 import SettingsIcon from 'mdi-react/SettingsIcon'
-import TuneVerticalIcon from 'mdi-react/TuneVerticalIcon'
 import { UserAreaHeaderNavItem } from './UserAreaHeader'
 
 export const userAreaHeaderNavItems: ReadonlyArray<UserAreaHeaderNavItem> = [
@@ -10,15 +9,8 @@ export const userAreaHeaderNavItems: ReadonlyArray<UserAreaHeaderNavItem> = [
     },
     {
         to: '/settings',
-        exact: true,
         label: 'Settings',
         icon: SettingsIcon,
-        condition: ({ user: { viewerCanAdminister } }) => viewerCanAdminister,
-    },
-    {
-        to: '/account',
-        label: 'Account',
-        icon: TuneVerticalIcon,
         condition: ({ user: { viewerCanAdminister } }) => viewerCanAdminister,
     },
 ]
